backend/pyproject.toml
[tool.poetry]
name = "ai-cfo-suite-phoenix"
version = "3.1.0"
description = "AI CFO Suite Phoenix - Multi-agent financial analysis system with RAG"
authors = ["Phoenix Team <team@aicfo.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"

# FastAPI and server
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"

# Vector DB
qdrant-client = "^1.7.3"

# AI Agents
crewai = "^0.11.0"
autogen-agentchat = "^0.2.36"

# LLM
openai = "^1.10.0"

# Cache and Storage
redis = "^5.0.1"
minio = "^7.2.3"

# SSH for remote agents
paramiko = "^3.4.0"

# Document Processing
pypdf = "^4.0.1"
python-docx = "^1.1.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"

# Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# Utilities
python-dateutil = "^2.8.2"
httpx = "^0.26.0"
tenacity = "^8.2.3"
requests = "^2.31.0"

# Monitoring and Logging
prometheus-client = "^0.19.0"
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.poetry.group.embedding]
# Optional group for embedding functionality
optional = true

[tool.poetry.group.embedding.dependencies]
llama-index = "^0.10.68"
llama-index-vector-stores-qdrant = "^0.2.11"
sentence-transformers = "^2.3.1"
huggingface-hub = "^0.20.3"
transformers = "^4.37.2"
torch = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

